REFACTOR Command line client to provide it as a library
