+++ lib/errors.js
@@ -142,7 +142,6 @@ module.exports = {
         } else {
             this.message = 'Couldn\t find device group';
         }
-        this.code = 400;
+        this.code = 404;
     }
-
 };
