+++ lib/fiware-iotagent-lib.js
@@ -164,6 +164,8 @@ exports.resetMiddlewares = ngsi.resetMiddlewares;
 exports.statsRegistry = statsRegistry;
 exports.clearRegistry = deviceService.clearRegistry;
 
+exports.commandLine = require('./command/commandLine');
+
 exports.dataPlugins = {
     compressTimestamp: require('./plugins/compressTimestamp'),
     attributeAlias: require('./plugins/attributeAlias')
